Make timidity to compile using cplusplus compilers: add missing type casts, avoid `new' as variable name, assign pointers to NULL instead of 0. Change type of next from void* to a pointer of structure's own type in PathList and MidEventList.
git-svn-id: file:///svn/uhexen2/code/trunk@3829 eaf649e6-6c13-0410-b194-84221022858e

+++ libs/timidity/instrum_dls.c
@@ -80,7 +80,7 @@ static int ChunkHasSubType(uint32 magic)
         RIFF, LIST
     };
     int i;
-    for ( i = 0; i < (sizeof(chunk_list) / sizeof(uint32)); ++i ) {
+    for ( i = 0; i < (int)(sizeof(chunk_list) / sizeof(uint32)); ++i ) {
         if ( magic == chunk_list[i] ) {
             return 1;
         }
@@ -94,7 +94,7 @@ static int ChunkHasSubChunks(uint32 magic)
         RIFF, LIST
     };
     int i;
-    for ( i = 0; i < (sizeof(chunk_list) / sizeof(uint32)); ++i ) {
+    for ( i = 0; i < (int)(sizeof(chunk_list) / sizeof(uint32)); ++i ) {
         if ( magic == chunk_list[i] ) {
             return 1;
         }
